Function cairo_rs::cairo_run::write_binary_trace

pub fn write_binary_trace(
    relocated_trace: &[RelocatedTraceEntry],
    trace_file: &Path
) -> Result<()>

Writes a trace as a binary file. Bincode encodes to little endian by default and each trace entry is composed of 3 usize values that are padded to always reach 64 bit size.