Module cairo_rs::math_utils

Functions

as_int

Returns the lift of the given field element, val, as an integer in the range (-prime/2, prime/2).

div_mod

Finds a nonnegative integer x < p such that (m * x) % p == n.

ec_add

Gets two points on an elliptic curve mod p and returns their sum. Assumes the points are given in affine form (x, y) and have different x coordinates.

ec_double

Doubles a point on an elliptic curve with the equation y^2 = x^3 + alpha*x + beta mod p. Assumes the point is given in affine form (x, y) and has y != 0.

ec_double_slope

Computes the slope of an elliptic curve with the equation y^2 = x^3 + alpha*x + beta mod p, at the given point. Assumes the point is given in affine form (x, y) and has y != 0.

isqrt

Returns the integer square root of the nonnegative integer n. This is the floor of the exact square root of n. Unlike math.sqrt(), this function doesn’t have rounding error issues.

line_slope

Computes the slope of the line connecting the two given EC points over the field GF(p). Assumes the points are given in affine form (x, y) and have different x coordinates.

safe_div

Performs integer division between x and y; fails if x is not divisible by y.