Function cairo_rs::math_utils::ec_double

pub fn ec_double(
    point: (BigInt, BigInt),
    alpha: &BigInt,
    prime: &BigInt
) -> (BigInt, BigInt)

Doubles a point on an elliptic curve with the equation y^2 = x^3 + alpha*x + beta mod p. Assumes the point is given in affine form (x, y) and has y != 0.