Trait cairo_rs::vm::runners::builtin_runner::BuiltinRunner

pub trait BuiltinRunner {
    fn initialize_segments(
        &mut self,
        segments: &mut MemorySegmentManager,
        memory: &mut Memory
    );
    fn initial_stack(&self) -> Vec<MaybeRelocatable>;
    fn base(&self) -> isize;
    fn add_validation_rule(&self, memory: &mut Memory) -> Result<(), RunnerError>;
    fn deduce_memory_cell(
        &mut self,
        address: &Relocatable,
        memory: &Memory
    ) -> Result<Option<MaybeRelocatable>, RunnerError>;
    fn as_any(&self) -> &dyn Any;

    fn get_memory_accesses(
        &self,
        vm: &VirtualMachine
    ) -> Result<Vec<Relocatable>, MemoryError> { ... }
}

Required Methods

fn initialize_segments(
&mut self,
segments: &mut MemorySegmentManager,
memory: &mut Memory
)

Creates the necessary segments for the builtin in the MemorySegmentManager and stores the first address on the builtin’s base

fn initial_stack(&self) -> Vec<MaybeRelocatable>

fn base(&self) -> isize

Returns the builtin’s base segment (offset is always zero).

fn add_validation_rule(&self, memory: &mut Memory) -> Result<(), RunnerError>

fn deduce_memory_cell(
&mut self,
address: &Relocatable,
memory: &Memory
) -> Result<Option<MaybeRelocatable>, RunnerError>

fn as_any(&self) -> &dyn Any

Provided Methods

fn get_memory_accesses(
&self,
vm: &VirtualMachine
) -> Result<Vec<Relocatable>, MemoryError>