Struct cairo_rs::vm::runners::cairo_runner::CairoRunner

pub struct CairoRunner {
    pub relocated_memory: Vec<Option<BigInt>>,
    pub relocated_trace: Option<Vec<RelocatedTraceEntry>>,
    pub exec_scopes: ExecutionScopes,
    /* private fields */
}

Fields
relocated_memory: Vec<Option<BigInt>>relocated_trace: Option<Vec<RelocatedTraceEntry>>exec_scopes: ExecutionScopes
Implementations

impl CairoRunner

pub fn new(program: &Program) -> Result<CairoRunner, RunnerError>

pub fn initialize(
&mut self,
vm: &mut VirtualMachine
) -> Result<Relocatable, RunnerError>

pub fn get_reference_list(&self) -> HashMap<usize, HintReference>

pub fn get_hint_data_dictionary(
&self,
references: &HashMap<usize, HintReference>,
hint_executor: &dyn HintProcessor
) -> Result<HashMap<usize, Vec<Box<dyn Any>>>, VirtualMachineError>

Gets the data used by the HintProcessor to execute each hint

pub fn get_constants(&self) -> &HashMap<String, BigInt>

pub fn run_until_pc(
&mut self,
address: Relocatable,
vm: &mut VirtualMachine,
hint_processor: &dyn HintProcessor
) -> Result<(), VirtualMachineError>

pub fn mark_as_accessed(
&mut self,
address: Relocatable,
size: usize
) -> Result<(), VirtualMachineError>

Mark a memory address as accesed.

pub fn get_memory_holes(&self, vm: &VirtualMachine) -> Result<usize, MemoryError>

Count the number of holes present in the segments.

pub fn relocate(&mut self, vm: &mut VirtualMachine) -> Result<(), TraceError>

pub fn get_output(
&mut self,
vm: &mut VirtualMachine
) -> Result<Option<String>, RunnerError>

pub fn write_output(
&mut self,
vm: &mut VirtualMachine,
stdout: &mut dyn Write
) -> Result<(), RunnerError>

Writes the values hosted in the output builtin’s segment Does nothing if the output builtin is not present in the program