Struct cairo_rs::vm::runners::cairo_runner::CairoRunner

pub struct CairoRunner {
    pub original_steps: Option<usize>,
    pub relocated_memory: Vec<Option<BigInt>>,
    pub relocated_trace: Option<Vec<RelocatedTraceEntry>>,
    pub exec_scopes: ExecutionScopes,
    /* private fields */
}

Fields
original_steps: Option<usize>relocated_memory: Vec<Option<BigInt>>relocated_trace: Option<Vec<RelocatedTraceEntry>>exec_scopes: ExecutionScopes
Implementations

impl CairoRunner

pub fn new(
program: &Program,
layout: &str,
proof_mode: bool
) -> Result<CairoRunner, RunnerError>

pub fn initialize(
&mut self,
vm: &mut VirtualMachine
) -> Result<Relocatable, RunnerError>

pub fn initialize_builtins(
&self,
vm: &mut VirtualMachine
) -> Result<(), RunnerError>

pub fn initialize_segments(
&mut self,
vm: &mut VirtualMachine,
program_base: Option<Relocatable>
)

Creates the necessary segments for the program, execution, and each builtin on the MemorySegmentManager and stores the first adress of each of this new segments as each owner’s base

pub fn initialize_function_entrypoint(
&mut self,
vm: &mut VirtualMachine,
entrypoint: usize,
stack: Vec<MaybeRelocatable>,
return_fp: MaybeRelocatable
) -> Result<Relocatable, RunnerError>

pub fn initialize_vm(
&mut self,
vm: &mut VirtualMachine
) -> Result<(), RunnerError>

pub fn get_initial_fp(&self) -> Option<Relocatable>

pub fn get_reference_list(&self) -> HashMap<usize, HintReference>

pub fn get_hint_data_dictionary(
&self,
references: &HashMap<usize, HintReference>,
hint_executor: &mut dyn HintProcessor
) -> Result<HashMap<usize, Vec<Box<dyn Any>>>, VirtualMachineError>

Gets the data used by the HintProcessor to execute each hint

pub fn get_constants(&self) -> &HashMap<String, BigInt>

pub fn get_program_builtins(&self) -> &Vec<String>

pub fn run_until_pc(
&mut self,
address: Relocatable,
vm: &mut VirtualMachine,
hint_processor: &mut dyn HintProcessor
) -> Result<(), VirtualMachineError>

pub fn run_for_steps(
&mut self,
steps: usize,
vm: &mut VirtualMachine,
hint_processor: &mut dyn HintProcessor
) -> Result<(), VirtualMachineError>

Execute an exact number of steps on the program from the actual position.

pub fn run_until_steps(
&mut self,
steps: usize,
vm: &mut VirtualMachine,
hint_processor: &mut dyn HintProcessor
) -> Result<(), VirtualMachineError>

Execute steps until a number of steps since the start of the program is reached.

pub fn run_until_next_power_of_2(
&mut self,
vm: &mut VirtualMachine,
hint_processor: &mut dyn HintProcessor
) -> Result<(), VirtualMachineError>

Execute steps until the step counter reaches a power of two.

pub fn get_perm_range_check_limits(
&self,
vm: &VirtualMachine
) -> Result<Option<(isize, isize)>, VirtualMachineError>

pub fn check_range_check_usage(
&self,
vm: &VirtualMachine
) -> Result<(), VirtualMachineError>

Checks that there are enough trace cells to fill the entire range check range.

pub fn get_memory_holes(&self, vm: &VirtualMachine) -> Result<usize, MemoryError>

Count the number of holes present in the segments.

pub fn check_diluted_check_usage(
&self,
vm: &VirtualMachine
) -> Result<(), VirtualMachineError>

Check if there are enough trace cells to fill the entire diluted checks.

pub fn end_run(
&mut self,
disable_trace_padding: bool,
disable_finalize_all: bool,
vm: &mut VirtualMachine,
hint_processor: &mut dyn HintProcessor
) -> Result<(), VirtualMachineError>

pub fn relocate(&mut self, vm: &mut VirtualMachine) -> Result<(), TraceError>

pub fn get_builtin_segments_info(
&self,
vm: &VirtualMachine
) -> Result<HashMap<&'static str, SegmentInfo>, RunnerError>

pub fn get_execution_resources(
&self,
vm: &VirtualMachine
) -> Result<ExecutionResources, TraceError>

pub fn get_output(
&mut self,
vm: &mut VirtualMachine
) -> Result<String, RunnerError>

pub fn write_output(
&mut self,
vm: &mut VirtualMachine,
stdout: &mut dyn Write
) -> Result<(), RunnerError>

Writes the values hosted in the output builtin’s segment. Does nothing if the output builtin is not present in the program.

pub fn finalize_segments(
&mut self,
vm: &mut VirtualMachine
) -> Result<(), RunnerError>

pub fn run_from_entrypoint(
&mut self,
entrypoint: usize,
args: Vec<&dyn Any>,
typed_args: bool,
verify_secure: bool,
apply_modulo_to_args: bool,
vm: &mut VirtualMachine,
hint_processor: &mut dyn HintProcessor
) -> Result<(), VirtualMachineError>

pub fn check_used_cells(
&self,
vm: &VirtualMachine
) -> Result<(), VirtualMachineError>

pub fn check_memory_usage(
&self,
vm: &VirtualMachine
) -> Result<(), VirtualMachineError>

pub fn initialize_function_runner(
&mut self,
vm: &mut VirtualMachine
) -> Result<(), RunnerError>

pub fn set_entrypoint(
&mut self,
new_entrypoint: Option<&str>
) -> Result<(), ProgramError>

Overrides the previous entrypoint with a custom one, or “main” if none is specified.

pub fn read_return_values(
&mut self,
vm: &VirtualMachine
) -> Result<(), RunnerError>

pub fn add_additional_hash_builtin(&self, vm: &mut VirtualMachine) -> Relocatable

Add (or replace if already present) a custom hash builtin. Returns a Relocatable with the new builtin base as the segment index.